Split up __main__.py
- The CLI is now defined in cli.py
- The different actions that deployer can perform, as selected by CLI
  subcommands have been moved to deploy_actions.py
- __main__.py now purely acts as an entrypoint for the deployer package
